deps/v8/test/webkit/fast/js/Promise-init-callback-receiver-expected.txt

|this| in Promise constructor should be undefined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS receiverInStrict is undefined
PASS successfullyParsed is true

TEST COMPLETE
